Implement TrafficLightSimulator
Added initial implementation of TrafficLightSimulator 
this code show the rules of the traffic signals ,when the red colored light is on then it indicates  that to stop the vechile for few seconds/minutes. orange colored lights indicates that be ready to go. Green colored light indicates that go from the ready state.